+++ lib/core/src/display.rs
@@ -2,12 +2,13 @@
 //! rendering utilities, etc.
 
 pub mod camera;
+pub mod navigation;
+pub mod object;
+pub mod render;
+pub mod scene;
 pub mod shape;
 pub mod symbol;
-pub mod object;
-pub mod scene;
 pub mod world;
-pub mod navigation;
