+++ lib/core/src/display/render/passes.rs
@@ -1,0 +1,16 @@
+//! Root module for render passes definitions.
+
+pub mod display_object;
+pub mod pixel_read;
+pub mod screen;
+
+
+
+/// Common types.
+pub mod types {
+    use super::*;
+    pub use display_object::*;
+    pub use pixel_read::*;
+    pub use screen::*;
+}
+pub use types::*;
