+++ lib/core/src/system/gpu/data.rs
@@ -28,16 +28,18 @@ pub use uniform::*;
 pub mod types {
     use super::*;
     pub use attribute::Attribute;
-    pub use attribute::AttributeScope;
     pub use attribute::AttributeInstanceIndex;
+    pub use attribute::AttributeScope;
     pub use buffer::AnyBuffer;
     pub use buffer::Buffer;
-    pub use buffer::IsBuffer;
     pub use buffer::BufferItem;
+    pub use buffer::IsBuffer;
     pub use default::GpuDefault;
     pub use gl_enum::GlEnum;
     pub use gl_enum::traits::*;
     pub use prim::*;
+    pub use super::texture;
+    pub use texture::Texture;
     pub use texture::TextureBindGuard;
     pub use uniform::AnyPrimUniform;
     pub use uniform::AnyTextureUniform;
